Put all CP Macro tests together
There were 2 QUnit.module('CP macros') in different files
